pyproject.toml
[project]
name = "chat-with-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.8",
    "langchain-openai>=1.1.7",
    "numpy>=2.4.2",
    "python-dotenv>=1.2.1",
    "sqlite-vec>=0.1.6",
]
